+++ src/main/java/cs3822/ModelStorage.java
@@ -8,7 +8,7 @@ interface ModelStorage {
   public void insert(int hash, Action action, float reward);
   public Action fetchAction(int hash);
   public float fetchReward(int hash);
-  public int getElemCount();
+  public long getElemCount();
   public boolean contains(int hash);
 
 }
